initial combine code

+++ neighbor.py
@@ -38,7 +38,7 @@ class Neighbor:
         self.val_tags = val["tags"].copy()
         self.val_updt_date = val["updt_date"].copy()
 
-        self.song_meta_issue_date = song_meta["issue_date"].copy()
+        self.song_meta_issue_date = song_meta["issue_date"].copy().astype(np.int64)
 
 
         ### ?. parameters
